only one source of truth for auth module info
